Initialized an R package. I wrote a number of funtions core to Nanoblot.
There are 2 helper functions: isUnique() and extractNanoblotData() which are requied for lapply-ing a few things.

scanBamFiles() is a command which 1) checks for uniqueness of SampleIDs and BamFiles 2) Runs Rsamtools::scanBam() and returns the output as a list of the outputs.

Because of how this was implimented it creates a list with n objects of length 1. Each of these objects is a list of length 13 (the output of scanBams). This could be considered a bug but I'm intentially leaving it so that additional data can be added to the list without modifing the standard output of scanBams().

bamFilesToNanoblotData() takes the output of  scanBamFiles() and returns a dataframe with the data required to make a NanoBlot. It uses the extractNanoblotData() helper function.

Finally, makeNanoblot() takes the output of bamFilesToNanoblotData(), a metadata file, a type of blot and and optional plotTitle. It prints the Nanoblot.